pyproject.toml
[project]
name = "golfcourseapi-scraper"
version = "0.1.0"
description = "A containerized scraper for the Golf Course API that systematically collects golf course data"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]
